Simple Refractoring
